requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
APScheduler==3.10.1
async-timeout==4.0.2
attrs==22.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.2
bilibili-api-python==15.3.1
Brotli==1.0.9
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
edge-tts==6.1.3
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
lxml==4.9.2
multidict==6.0.4
openai==0.27.2
Pillow==9.4.0
pyasn1==0.4.8
pygame==2.3.0
pypng==0.20220715.0
pytz==2023.2
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.4.2
requests==2.28.2
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.2
tzlocal==4.3
urllib3==1.26.15
yarl==1.8.2
